So many changes. 1. Category calculations 2. Changeable Budget start day 3. Some icons, migrations and more. Check diff for more details.

+++ database/migrations/2017_04_30_080400_create_subcategories_table.php
@@ -18,7 +18,7 @@ class CreateSubcategoriesTable extends Migration
             $table->integer('category_id')->index()->unsigned();
             $table->integer('user_id')->index();
             $table->string('subcategoryName');
-            $table->integer('subcategoryBudget');
+            $table->decimal('subcategoryBudget', 20, 2);
             $table->softDeletes();
             $table->timestamps();
